Rename shared code files
