+++ src/ip/client.c
@@ -1,5 +1,5 @@
-#include "client_utils.h"
-#include "utils.h"
+#include "ip_client_utils.h"
+#include "ip_socket.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
