+++ src/ip/ip_client_utils.c
@@ -1,5 +1,5 @@
+#include "ip_socket.h"
 #include "types.h"
-#include "utils.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
